# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentio SMTP API",
    "description": "Multi-tenant SMTP platform API",
    "contact": {
      "name": "Truespar Team"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/tracking-domains": {
      "get": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "list_tracking_domains",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_TrackingDomainResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "create_tracking_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrackingDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TrackingDomainResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/tracking-domains/{id}": {
      "get": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "get_tracking_domain",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TrackingDomainResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "delete_tracking_domain",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tracking domain deleted"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/tracking-domains/{id}/certificate": {
      "get": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "get_certificate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TrackingCertificateResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "upload_certificate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCertificateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TrackingCertificateResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/tracking-domains/{id}/verify": {
      "post": {
        "tags": [
          "Tracking Domains"
        ],
        "operationId": "verify_tracking_domain",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TrackingDomainResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCertificateRequest": {
        "type": "object",
        "required": [
          "certificate",
          "private_key",
          "expires_at",
          "renew_after"
        ],
        "properties": {
          "certificate": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "intermediaries": {
            "type": [
              "string",
              "null"
            ]
          },
          "private_key": {
            "type": "string"
          },
          "renew_after": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateTrackingDomainRequest": {
        "type": "object",
        "required": [
          "domain_name",
          "cname_target"
        ],
        "properties": {
          "cname_target": {
            "type": "string"
          },
          "domain_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "domain_name": {
            "type": "string"
          },
          "ssl_enabled": {
            "type": "boolean"
          },
          "track_clicks": {
            "type": "boolean"
          },
          "track_opens": {
            "type": "boolean"
          }
        }
      },
...
      "DataResponse_TrackingCertificateResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "tracking_domain_id",
              "expires_at",
              "renew_after",
              "created_at"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "id": {
                "$ref": "#/components/schemas/TrackingCertificateId"
              },
              "renew_after": {
                "type": "string",
                "format": "date-time"
              },
              "tracking_domain_id": {
                "$ref": "#/components/schemas/TrackingDomainId"
              }
            }
          }
        }
      },
      "DataResponse_TrackingDomainResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "domain_name",
              "cname_target",
              "dns_status",
              "ssl_enabled",
              "track_opens",
              "track_clicks",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "cname_target": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "dns_checked_at": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "dns_error": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "dns_status": {
                "$ref": "#/components/schemas/DomainStatus"
              },
              "domain_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/DomainId"
                  }
                ]
              },
              "domain_name": {
                "type": "string"
              },
              "id": {
                "$ref": "#/components/schemas/TrackingDomainId"
              },
              "ssl_enabled": {
                "type": "boolean"
              },
              "track_clicks": {
                "type": "boolean"
              },
              "track_opens": {
                "type": "boolean"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
...
      "DataResponse_Vec_TrackingDomainResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "domain_name",
                "cname_target",
                "dns_status",
                "ssl_enabled",
                "track_opens",
                "track_clicks",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "cname_target": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "dns_checked_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "dns_error": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "dns_status": {
                  "$ref": "#/components/schemas/DomainStatus"
                },
                "domain_id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/DomainId"
                    }
                  ]
                },
                "domain_name": {
                  "type": "string"
                },
                "id": {
                  "$ref": "#/components/schemas/TrackingDomainId"
                },
                "ssl_enabled": {
                  "type": "boolean"
                },
                "track_clicks": {
                  "type": "boolean"
                },
                "track_opens": {
                  "type": "boolean"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "DomainId": {
        "type": "string",
        "format": "uuid"
      },
...
      "DomainStatus": {
        "type": "string",
        "enum": [
          "pending",
          "verified",
          "failed"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "TrackingCertificateId": {
        "type": "string",
        "format": "uuid"
      },
...
      "TrackingDomainId": {
        "type": "string",
        "format": "uuid"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
...
}